curve25519_dalek::backend

Module serial

Serial implementations of field, scalar, point arithmetic.

When the vector backend is disabled, the crate uses the mixed-model strategy for implementing point operations and scalar multiplication; see the curve_models and scalar_mul documentation for more information.

When the vector backend is enabled, the field and scalar implementations are still used for non-vectorized operations.

Modules

curve_models

Internal curve representations which are not part of the public API.

scalar_mul

Implementations of various scalar multiplication algorithms.