curve25519_dalek::backend::serial

Module scalar_mul

Implementations of various scalar multiplication algorithms.

Note that all of these implementations use serial code for field arithmetic with the multi-model strategy described in the curve_models module. The vectorized AVX2 backend has its own scalar multiplication implementations, since it only uses one curve model.

Modules

pippenger

Implementation of a variant of Pippenger’s algorithm.

precomputed_straus

Precomputation for Straus’s method.

straus

Implementation of the interleaved window method, also known as Straus’ method.

variable_base

vartime_double_base