curve25519_dalek::backend::serial::scalar_mul
Module straus

Available on crate feature alloc only.

Implementation of the interleaved window method, also known as Straus’ method.

Structs

Straus

Perform multiscalar multiplication by the interleaved window method, also known as Straus’ method (since it was apparently first published by Straus in 1964, as a solution to a problem posted in the American Mathematical Monthly in 1963).