Struct curve25519_dalek::backend::vector::scalar_mul::straus::spec_avx2::Straus

pub struct Straus {}

Available on curve25519_dalek_backend="simd" and crate feature alloc only.

Multiscalar multiplication using interleaved window / Straus’ method. See the Straus struct in the serial backend for more details.

This exists as a seperate implementation from that one because the AVX2 code uses different curve models (it does not pass between multiple models during scalar mul), and it has to convert the point representation on the fly.