curve25519_dalek::edwards

Struct EdwardsPoint

pub struct EdwardsPoint {
    pub(crate) X: FieldElement51,
    pub(crate) Y: FieldElement51,
    pub(crate) Z: FieldElement51,
    pub(crate) T: FieldElement51,
}

An EdwardsPoint represents a point on the Edwards form of Curve25519.

Fields
X: FieldElement51Y: FieldElement51Z: FieldElement51T: FieldElement51
Implementations

impl EdwardsPoint

pub(crate) fn as_projective_niels(&self) -> ProjectiveNielsPoint

Convert to a ProjectiveNielsPoint

pub(crate) const fn as_projective(&self) -> ProjectivePoint

Convert the representation of this point from extended coordinates to projective coordinates.

Free.

pub(crate) fn as_affine_niels(&self) -> AffineNielsPoint

Dehomogenize to a AffineNielsPoint. Mainly for testing.

pub(crate) fn to_affine(self) -> AffinePoint

Dehomogenize to AffinePoint.

pub fn to_montgomery(&self) -> MontgomeryPoint

Convert this EdwardsPoint on the Edwards model to the corresponding MontgomeryPoint on the Montgomery model.

This function has one exceptional case; the identity point of the Edwards curve is sent to the 2-torsion point \((0,0)\) on the Montgomery curve.

Note that this is a one-way conversion, since the Montgomery model does not retain sign information.

pub fn to_montgomery_batch(eds: &[Self]) -> Vec<MontgomeryPoint>

Converts a large batch of points to Edwards at once. This has the same behavior on identity elements as Self::to_montgomery.

pub fn compress(&self) -> CompressedEdwardsY

Compress this point to CompressedEdwardsY format.

pub fn compress_batch<const N: usize>( inputs: &[EdwardsPoint; N], ) -> [CompressedEdwardsY; N]

Compress several EdwardsPoints into CompressedEdwardsY format, using a batch inversion for a significant speedup.

pub fn compress_batch_alloc(inputs: &[EdwardsPoint]) -> Vec<CompressedEdwardsY>

Compress several EdwardsPoints into CompressedEdwardsY format, using a batch inversion for a significant speedup.

fn map_to_curve(fe: FieldElement51) -> EdwardsPoint

pub fn encode_to_curve<D>(bytes: &[&[u8]], domain_sep: &[&[u8]]) -> EdwardsPoint
where D: BlockSizeUser + Default + FixedOutput<OutputSize = U64> + HashMarker, D::BlockSize: IsGreater<D::OutputSize, Output = True>,

Perform encode to curve per RFC 9380, with explicit hash function and domain separator domain_sep, using the Twisted Edwards Elligator 2 method. The input is the concatenation of the elements of bytes. Likewise for the domain separator with domain_sep. At least one element of domain_sep, MUST be nonempty, and the concatenation MUST NOT exceed 255 bytes.

The specification names SHA-512 as an example of a secure hash to use with this function, but you may use any 512-bit hash within reason (see the spec for details).

Warning

encode_to_curve is a nonuniform encoding from byte strings to points in G. That is, the distribution of its output is not uniformly random in G: the set of possible outputs of encode_to_curve is only a fraction of the points in G, and some points in this set are more likely to be output than others.

If your application needs the distribution of the output to be statistically close to uniform in G, use Self::hash_to_curve instead.

Panics

Panics if domain_sep.collect().len() == 0 or > 255

pub fn hash_to_curve<D>(bytes: &[&[u8]], domain_sep: &[&[u8]]) -> EdwardsPoint
where D: BlockSizeUser + Default + FixedOutput<OutputSize = U64> + HashMarker, D::BlockSize: IsGreater<D::OutputSize, Output = True>,

Perform a hash to curve per RFC 9380, with explicit hash function and domain separator domain_sep, using the Twisted Edwards Elligator 2 method. The input is the concatenation of the elements of bytes. Likewise for the domain separator with domain_sep. At least one element of domain_sep, MUST be nonempty, and the concatenation MUST NOT exceed 255 bytes.

The specification names SHA-512 as an example of a secure hash to use with this function, but you may use any 512-bit hash within reason (see the spec for details).

Panics

Panics if domain_sep.collect().len() == 0 or > 255

pub fn random<R: Rng + ?Sized>(rng: &mut R) -> Self

Return an EdwardsPoint chosen uniformly at random using a user-provided RNG.

Inputs

rng: any RNG which implements Rng

Returns

A random EdwardsPoint.

Implementation

Uses rejection sampling, generating a random CompressedEdwardsY and then attempting point decompression, rejecting invalid points.

impl EdwardsPoint

pub(crate) fn double(&self) -> EdwardsPoint

Add this point to itself.

impl EdwardsPoint

pub fn mul_base(scalar: &Scalar) -> Self

Fixed-base scalar multiplication by the Ed25519 base point.

Uses precomputed basepoint tables when the precomputed-tables feature is enabled, trading off increased code size for ~4x better performance.

pub fn mul_clamped(self, bytes: [u8; 32]) -> Self

Multiply this point by clamp_integer(bytes). For a description of clamping, see clamp_integer.

pub fn mul_base_clamped(bytes: [u8; 32]) -> Self

Multiply the basepoint by clamp_integer(bytes). For a description of clamping, see clamp_integer.

impl EdwardsPoint

pub fn vartime_double_scalar_mul_basepoint( a: &Scalar, A: &EdwardsPoint, b: &Scalar, ) -> EdwardsPoint

Compute \(aA + bB\) in variable time, where \(B\) is the Ed25519 basepoint.

impl EdwardsPoint

pub fn mul_by_cofactor(&self) -> EdwardsPoint

Multiply by the cofactor: return \([8]P\).

pub(crate) fn mul_by_pow_2(&self, k: u32) -> EdwardsPoint

Compute \([2^k] P \) by successive doublings. Requires \( k > 0 \).

pub fn is_small_order(&self) -> bool

Determine if this point is of small order.

Return

true if self is in the torsion subgroup \( \mathcal E[8] \);

false if self is not in the torsion subgroup \( \mathcal E[8] \).

Example

use curve25519_dalek::constants;

// Generator of the prime-order subgroup
let P = constants::ED25519_BASEPOINT_POINT;
// Generator of the torsion subgroup
let Q = constants::EIGHT_TORSION[1];

// P has large order
assert_eq!(P.is_small_order(), false);

// Q has small order
assert_eq!(Q.is_small_order(), true);

pub fn is_torsion_free(&self) -> bool

Determine if this point is “torsion-free”, i.e., is contained in the prime-order subgroup.

Return

true if self has zero torsion component and is in the prime-order subgroup;

false if self has a nonzero torsion component and is not in the prime-order subgroup.

Example

use curve25519_dalek::constants;

// Generator of the prime-order subgroup
let P = constants::ED25519_BASEPOINT_POINT;
// Generator of the torsion subgroup
let Q = constants::EIGHT_TORSION[1];

// P is torsion-free
assert_eq!(P.is_torsion_free(), true);

// P + Q is not torsion-free
assert_eq!((P+Q).is_torsion_free(), false);