Struct curve25519_dalek::edwards::ExtendedPoint

pub struct ExtendedPoint { /* fields omitted */ }

An ExtendedPoint is a point on the curve in 𝗣³(𝔽ₚ). A point (x,y) in the affine model corresponds to (x:y:1:xy).

Methods

impl ExtendedPoint

pub fn to_montgomery(&self) -> MontgomeryPoint

Convert this ExtendedPoint on the Edwards model to the corresponding MontgomeryPoint on the Montgomery model.

Note that this is a one-way conversion, since the Montgomery model does not retain sign information.

pub fn compress(&self) -> CompressedEdwardsY

Compress this point to CompressedEdwardsY format.

impl ExtendedPoint

pub fn mult_by_cofactor(&self) -> ExtendedPoint

Multiply by the cofactor: compute 8 * self.

pub fn is_small_order(&self) -> bool

Determine if this point is of small order.

The order of the group of points on the curve Ɛ is |Ɛ| = 8q. Thus, to check if a point P is of small order, we multiply by 8 and then test if the result is equal to the identity.

Return

True if it is of small order; false otherwise.