Struct curve25519_dalek::field::FieldElement51

pub struct FieldElement51(_);

A FieldElement51 represents an element of the field \( \mathbb Z / (2^{255} - 19)\).

In the 64-bit implementation, a FieldElement is represented in radix \(2^{51}\) as five u64s; the coefficients are allowed to grow up to \(2^{54}\) between reductions modulo \(p\).

Note

The curve25519_dalek::field module provides a type alias curve25519_dalek::field::FieldElement to either FieldElement51 or FieldElement2625.

The backend-specific type FieldElement51 should not be used outside of the curve25519_dalek::field module.

Implementations

impl FieldElement51

pub fn is_negative(&self) -> Choice

Determine if this FieldElement is negative, in the sense used in the ed25519 paper: x is negative if the low bit is set.

Return

If negative, return Choice(1). Otherwise, return Choice(0).

pub fn is_zero(&self) -> Choice

Determine if this FieldElement is zero.

Return

If zero, return Choice(1). Otherwise, return Choice(0).

pub fn batch_invert(inputs: &mut [FieldElement])

Given a slice of public FieldElements, replace each with its inverse.

All input FieldElements MUST be nonzero.

pub fn invert(&self) -> FieldElement

Given a nonzero field element, compute its inverse.

The inverse is computed as self^(p-2), since x^(p-2)x = x^(p-1) = 1 (mod p).

This function returns zero on input zero.

pub fn sqrt_ratio_i(
u: &FieldElement,
v: &FieldElement
) -> (Choice, FieldElement)

Given FieldElements u and v, compute either sqrt(u/v) or sqrt(i*u/v) in constant time.

This function always returns the nonnegative square root.

Return

(Choice(1), +sqrt(u/v)) if v is nonzero and u/v is square;

(Choice(1), zero) if u is zero;

(Choice(0), zero) if v is zero and u is nonzero;

(Choice(0), +sqrt(i*u/v)) if u/v is nonsquare (so i*u/v is square).

pub fn invsqrt(&self) -> (Choice, FieldElement)

Attempt to compute sqrt(1/self) in constant time.

Convenience wrapper around sqrt_ratio_i.

This function always returns the nonnegative square root.

Return

(Choice(1), +sqrt(1/self)) if self is a nonzero square;

(Choice(0), zero) if self is zero;

(Choice(0), +sqrt(i/self)) if self is a nonzero nonsquare;

impl FieldElement51

pub fn negate(&mut self)

Invert the sign of this field element

pub fn zero() -> FieldElement51

Construct zero.

pub fn one() -> FieldElement51

Construct one.

pub fn minus_one() -> FieldElement51

Construct -1.

pub fn from_bytes(bytes: &[u8; 32]) -> FieldElement51

Load a FieldElement51 from the low 255 bits of a 256-bit input.

Warning

This function does not check that the input used the canonical representative. It masks the high bit, but it will happily decode 2^255 - 18 to 1. Applications that require a canonical encoding of every field element should decode, re-encode to the canonical encoding, and check that the input was canonical.

pub fn to_bytes(&self) -> [u8; 32]

Serialize this FieldElement51 to a 32-byte array. The encoding is canonical.

pub fn pow2k(&self, k: u32) -> FieldElement51

Given k > 0, return self^(2^k).

pub fn square(&self) -> FieldElement51

Returns the square of this field element.

pub fn square2(&self) -> FieldElement51

Returns 2 times the square of this field element.