Struct curve25519_dalek::field::FieldElement64

pub struct FieldElement64(pub [u64; 5]);

A FieldElement64 represents an element of the field GF(2255 - 19).

In the 64-bit implementation, a FieldElement is represented in radix 251 as five u64s; the coefficients are allowed to grow up to 254 between reductions mod p.

Warning

You almost certainly do not want to use FieldElement64 directly. Consider using curve25519_dalek::field::FieldElement, which will automatically select between FieldElement32 and FieldElement64 depending on whether curve25519-dalek was compiled with --features="nightly".

This implementation, FieldElement64, is intended for x64_64 platforms, which have the MUL instructions taking 64-bit inputs and producing 128-bit outputs. On other platforms, this implementation is not recommended. On Haswell and newer, the BMI2 instruction set provides MULX and friends, which gives even better performance. This implementation requires Rust's u128, which is not yet stable.

Methods

impl FieldElement64

fn negate(&mut self)

Invert the sign of this field element

fn zero() -> FieldElement64

Construct zero.

fn one() -> FieldElement64

Construct one.

fn minus_one() -> FieldElement64

Construct -1.

fn from_bytes(bytes: &[u8; 32]) -> FieldElement64

Load a FieldElement64 from the low 255 bits of a 256-bit input.

Warning

This function does not check that the input used the canonical representative. It masks the high bit, but it will happily decode 2255 - 18 to 1. Applications that require a canonical encoding of every field element should decode, re-encode to the canonical encoding, and check that the input was canonical.

fn to_bytes(&self) -> [u8; 32]

Serialize this FieldElement64 to a 32-byte array. The encoding is canonical.

fn square(&self) -> FieldElement64

Returns the square of this field element.

fn square2(&self) -> FieldElement64

Returns 2 times the square of this field element.