Type Definition curve25519_dalek::field::FieldElement

type FieldElement = FieldElement51;

A FieldElement represents an element of the field \( \mathbb Z / (2^{255} - 19)\).

The FieldElement type is an alias for one of the platform-specific implementations.

Implementations

impl FieldElement

pub fn is_negative(&self) -> Choice

Determine if this FieldElement is negative, in the sense used in the ed25519 paper: x is negative if the low bit is set.

Return

If negative, return Choice(1). Otherwise, return Choice(0).

pub fn is_zero(&self) -> Choice

Determine if this FieldElement is zero.

Return

If zero, return Choice(1). Otherwise, return Choice(0).

pub fn batch_invert(inputs: &mut [FieldElement])

Given a slice of public FieldElements, replace each with its inverse.

All input FieldElements MUST be nonzero.

pub fn invert(&self) -> FieldElement

Given a nonzero field element, compute its inverse.

The inverse is computed as self^(p-2), since x^(p-2)x = x^(p-1) = 1 (mod p).

This function returns zero on input zero.

pub fn sqrt_ratio_i(
u: &FieldElement,
v: &FieldElement
) -> (Choice, FieldElement)

Given FieldElements u and v, compute either sqrt(u/v) or sqrt(i*u/v) in constant time.

This function always returns the nonnegative square root.

Return

(Choice(1), +sqrt(u/v)) if v is nonzero and u/v is square;

(Choice(1), zero) if u is zero;

(Choice(0), zero) if v is zero and u is nonzero;

(Choice(0), +sqrt(i*u/v)) if u/v is nonsquare (so i*u/v is square).

pub fn invsqrt(&self) -> (Choice, FieldElement)

Attempt to compute sqrt(1/self) in constant time.

Convenience wrapper around sqrt_ratio_i.

This function always returns the nonnegative square root.

Return

(Choice(1), +sqrt(1/self)) if self is a nonzero square;

(Choice(0), zero) if self is zero;

(Choice(0), +sqrt(i/self)) if self is a nonzero nonsquare;