Function curve25519_dalek::ristretto::multiscalar_mult

pub fn multiscalar_mult<'a, 'b, I, J>(scalars: I, points: J) -> RistrettoPoint where
    I: IntoIterator<Item = &'a Scalar>,
    J: IntoIterator<Item = &'b RistrettoPoint>, 

Given a vector of (possibly secret) scalars and a vector of (possibly secret) points, compute c_1 P_1 + ... + c_n P_n.

This function has the same behaviour as vartime::multiscalar_mult but is constant-time.

Input

An iterable of Scalars and a iterable of DecafPoints. It is an error to call this function with two iterators of different lengths.