curve25519_dalek::traits

Trait ValidityCheck

pub(crate) trait ValidityCheck {
    // Required method
    fn is_valid(&self) -> bool;
}

Trait for checking whether a point is on the curve.

This trait is only for debugging/testing, since it should be impossible for a curve25519-dalek user to construct an invalid point.

Required Methods

fn is_valid(&self) -> bool

Checks whether the point is on the curve. Not CT.