curve25519_dalek::backend::serial::curve_models

Struct AffineNielsPoint

pub struct AffineNielsPoint {
    pub y_plus_x: FieldElement51,
    pub y_minus_x: FieldElement51,
    pub xy2d: FieldElement51,
}

A pre-computed point in the affine model for the curve, represented as \((y+x, y-x, 2dxy)\) in โNiels coordinatesโ.

More details on the relationships between the different curve models can be found in the module-level documentation.

Fields
y_plus_x: FieldElement51y_minus_x: FieldElement51xy2d: FieldElement51