curve25519_dalek::backend::vector::avx2::field

Struct FieldElement2625x4

pub struct FieldElement2625x4(pub(crate) [u32x8; 5]);

A vector of four field elements.

Each operation on a FieldElement2625x4 has documented effects on the bounds of the coefficients. This API is designed for speed and not safety; it is the callerโs responsibility to ensure that the post-conditions of one operation are compatible with the pre-conditions of the next.

Tuple Fields
0: [u32x8; 5]
Implementations

impl FieldElement2625x4

pub const ZERO: FieldElement2625x4

pub fn split(&self) -> [FieldElement51; 4]

Split this vector into an array of four (serial) field elements.

pub fn shuffle(&self, control: Shuffle) -> FieldElement2625x4

Rearrange the elements of this vector according to control.

The control parameter should be a compile-time constant, so that when this function is inlined, LLVM is able to lower the shuffle using an immediate.

pub fn blend( &self, other: FieldElement2625x4, control: Lanes, ) -> FieldElement2625x4

Blend self with other, taking lanes specified in control from other.

The control parameter should be a compile-time constant, so that this function can be inlined and LLVM can lower it to a blend instruction using an immediate.

pub fn splat(x: &FieldElement51) -> FieldElement2625x4

Convenience wrapper around new(x,x,x,x).

pub fn new( x0: &FieldElement51, x1: &FieldElement51, x2: &FieldElement51, x3: &FieldElement51, ) -> FieldElement2625x4

Create a FieldElement2625x4 from four FieldElement51s.

Postconditions

The resulting FieldElement2625x4 is bounded with \( b < 0.0002 \).

pub fn negate_lazy(&self) -> FieldElement2625x4

Given \((A,B,C,D)\), compute \((-A,-B,-C,-D)\), without performing a reduction.

Preconditions

The coefficients of self must be bounded with \( b < 0.999 \).

Postconditions

The coefficients of the result are bounded with \( b < 1 \).

pub fn diff_sum(&self) -> FieldElement2625x4

Given self = (A,B,C,D), compute (B - A, B + A, D - C, D + C).

Preconditions

The coefficients of self must be bounded with \( b < 0.01 \).

Postconditions

The coefficients of the result are bounded with \( b < 1.6 \).

pub fn reduce(&self) -> FieldElement2625x4

Reduce this vector of field elements \(\mathrm{mod} p\).

Postconditions

The coefficients of the result are bounded with \( b < 0.0002 \).

fn reduce64(z: [u64x4; 10]) -> FieldElement2625x4

Given an array of wide coefficients, reduce them to a FieldElement2625x4.

Postconditions

The coefficients of the result are bounded with \( b < 0.007 \).

pub fn square_and_negate_D(&self) -> FieldElement2625x4

Square this field element, and negate the resultโs \(D\) value.

Preconditions

The coefficients of self must be bounded with \( b < 1.5 \).

Postconditions

The coefficients of the result are bounded with \( b < 0.007 \).