Struct curve25519_dalek::edwards::CompressedEdwardsY

pub struct CompressedEdwardsY(pub [u8; 32]);

In "Edwards y" format, the point (x,y) on the curve is determined by the y-coordinate and the sign of x, marshalled into a 32-byte array.

The first 255 bits of a CompressedEdwardsY represent the y-coordinate. The high bit of the 32nd byte gives the sign of x.

Methods

impl CompressedEdwardsY

pub fn as_bytes(&self) -> &[u8; 32]

View this CompressedEdwardsY as an array of bytes.

pub fn to_bytes(&self) -> [u8; 32]

Copy this CompressedEdwardsY to an array of bytes.

pub fn decompress(&self) -> Option<ExtendedPoint>

Attempt to decompress to an ExtendedPoint.

Returns None if the input is not the y-coordinate of a curve point.