Struct curve25519_dalek::edwards::EdwardsPoint

pub struct EdwardsPoint { /* fields omitted */ }

An EdwardsPoint represents a point on the Edwards form of Curve25519.

Implementations

impl EdwardsPoint

pub fn to_montgomery(&self) -> MontgomeryPoint

Convert this EdwardsPoint on the Edwards model to the corresponding MontgomeryPoint on the Montgomery model.

This function has one exceptional case; the identity point of the Edwards curve is sent to the 2-torsion point \((0,0)\) on the Montgomery curve.

Note that this is a one-way conversion, since the Montgomery model does not retain sign information.

pub fn compress(&self) -> CompressedEdwardsY

Compress this point to CompressedEdwardsY format.

pub fn hash_from_bytes<D>(bytes: &[u8]) -> EdwardsPoint where
D: Digest<OutputSize = U64> + Default,

Perform hashing to the group using the Elligator2 map

See https://tools.ietf.org/html/draft-irtf-cfrg-hash-to-curve-10#section-6.7.1

impl EdwardsPoint

pub fn vartime_double_scalar_mul_basepoint(
a: &Scalar,
A: &EdwardsPoint,
b: &Scalar
) -> EdwardsPoint

Compute \(aA + bB\) in variable time, where \(B\) is the Ed25519 basepoint.

impl EdwardsPoint

pub fn mul_by_cofactor(&self) -> EdwardsPoint

Multiply by the cofactor: return \([8]P\).

pub fn is_small_order(&self) -> bool

Determine if this point is of small order.

Return

true if self is in the torsion subgroup \( \mathcal E[8] \);

false if self is not in the torsion subgroup \( \mathcal E[8] \).

Example

use curve25519_dalek::constants;

// Generator of the prime-order subgroup
let P = constants::ED25519_BASEPOINT_POINT;
// Generator of the torsion subgroup
let Q = constants::EIGHT_TORSION[1];

// P has large order
assert_eq!(P.is_small_order(), false);

// Q has small order
assert_eq!(Q.is_small_order(), true);

pub fn is_torsion_free(&self) -> bool

Determine if this point is “torsion-free”, i.e., is contained in the prime-order subgroup.

Return

true if self has zero torsion component and is in the prime-order subgroup;

false if self has a nonzero torsion component and is not in the prime-order subgroup.

Example

use curve25519_dalek::constants;

// Generator of the prime-order subgroup
let P = constants::ED25519_BASEPOINT_POINT;
// Generator of the torsion subgroup
let Q = constants::EIGHT_TORSION[1];

// P is torsion-free
assert_eq!(P.is_torsion_free(), true);

// P + Q is not torsion-free
assert_eq!((P+Q).is_torsion_free(), false);