Module curve25519_dalek::field

Field arithmetic modulo \(p = 2^{255} - 19\).

The curve25519_dalek::field module provides a type alias curve25519_dalek::field::FieldElement to a field element type defined in the backend module; either FieldElement51 or FieldElement2625.

Field operations defined in terms of machine operations, such as field multiplication or squaring, are defined in the backend implementation.

Field operations defined in terms of other field operations, such as field inversion or square roots, are defined here.

Structs

FieldElement51
A FieldElement51 represents an element of the field \( \mathbb Z / (2^{255} - 19)\).

Type Definitions

FieldElement
A FieldElement represents an element of the field \( \mathbb Z / (2^{255} - 19)\).