Module curve25519_dalek::montgomery

Montgomery arithmetic.

Apart from the compressed point implementation (i.e. CompressedMontgomeryU), this module is a "clean room" implementation of the Montgomery arithmetic described in the following papers:

Costello, Craig, and Benjamin Smith. "Montgomery curves and their arithmetic." Journal of Cryptographic Engineering (2017): 1-14. PDF

Montgomery, Peter L. "Speeding the Pollard and elliptic curve methods of factorization." Mathematics of computation 48.177 (1987): 243-264. PDF

Structs

CompressedMontgomeryU
In "Montgomery u" format, as used in X25519, a point (u,v) on the Montgomery curve

MontgomeryPoint
A point on the Montgomery form of the curve, in projective 𝗣2 coordinates.