Struct curve25519_dalek::montgomery::MontgomeryPoint

pub struct MontgomeryPoint {
    pub U: FieldElement64,
    pub W: FieldElement64,
}

A point on the Montgomery form of the curve, in projective 𝗣2 coordinates.

The transition between affine and projective is given by

u → U/W v → V/W

thus the Montgomery curve equation

E_(A,B) : Bv² = u(u² + Au + 1)

becomes

E_(A,B) : BV²W = U(U² + AUW + W²) ⊆ 𝗣2

Here, again, to differentiate from points in the twisted Edwards model, we call the point (x,y) in affine coordinates (u,v) and similarly in projective space we use (U:V:W). However, since (as per Montgomery's original work) the v-coordinate is superfluous for the purposes of scalar multiplication, we merely use (U:W).

Fields
U: FieldElement64 W: FieldElement64
Methods

impl MontgomeryPoint

pub fn compress(&self) -> CompressedMontgomeryU

Compress this point to only its u-coordinate (note: affine).

Returns

A CompressedMontgomeryU.