Struct curve25519_dalek::montgomery::MontgomeryPoint

pub struct MontgomeryPoint(pub [u8; 32]);

Holds the \(u\)-coordinate of a point on the Montgomery form of Curve25519 or its twist.

Methods

impl MontgomeryPoint

pub fn as_bytes<'a>(&'a self) -> &'a [u8; 32]

View this MontgomeryPoint as an array of bytes.

pub fn to_bytes(&self) -> [u8; 32]

Convert this MontgomeryPoint to an array of bytes.

pub fn to_edwards(&self, sign: u8) -> Option<EdwardsPoint>

Attempt to convert to an EdwardsPoint, using the supplied choice of sign for the EdwardsPoint.

Return

Some(EdwardsPoint) if self is the \(u\)-coordinate of a point on (the Montgomery form of) Curve25519;

None if self is the \(u\)-coordinate of a point on the twist of (the Montgomery form of) Curve25519;