curve25519_dalek::montgomery
Struct MontgomeryPoint

pub struct MontgomeryPoint(pub [u8; 32]);

Holds the \(u\)-coordinate of a point on the Montgomery form of Curve25519 or its twist.

Tuple Fields
0: [u8; 32]
Implementations

impl MontgomeryPoint

pub fn mul_base(scalar: &Scalar) -> Self

Fixed-base scalar multiplication (i.e. multiplication by the base point).

pub fn mul_clamped(self, bytes: [u8; 32]) -> Self

Multiply this point by clamp_integer(bytes). For a description of clamping, see clamp_integer.

pub fn mul_base_clamped(bytes: [u8; 32]) -> Self

Multiply the basepoint by clamp_integer(bytes). For a description of clamping, see clamp_integer.

pub fn mul_bits_be(&self, bits: impl Iterator<Item = bool>) -> MontgomeryPoint

Given self \( = u_0(P) \), and a big-endian bit representation of an integer \(n\), return \( u_0([n]P) \). This is constant time in the length of bits.

NOTE: You probably do not want to use this function. Almost every protocol built on Curve25519 uses clamped multiplication, explained here. When in doubt, use Self::mul_clamped.

pub const fn as_bytes(&self) -> &[u8; 32]

View this MontgomeryPoint as an array of bytes.

pub const fn to_bytes(&self) -> [u8; 32]

Convert this MontgomeryPoint to an array of bytes.

pub fn to_edwards(&self, sign: u8) -> Option<EdwardsPoint>

Attempt to convert to an EdwardsPoint, using the supplied choice of sign for the EdwardsPoint.

Inputs

sign: a u8 donating the desired sign of the resulting EdwardsPoint. 0 denotes positive and 1 negative.

Return

Some(EdwardsPoint) if self is the \(u\)-coordinate of a point on (the Montgomery form of) Curve25519;

None if self is the \(u\)-coordinate of a point on the twist of (the Montgomery form of) Curve25519;