Struct curve25519_dalek::montgomery::MontgomeryPoint

pub struct MontgomeryPoint(pub [u8; 32]);

Holds the \(u\)-coordinate of a point on the Montgomery form of Curve25519 or its twist.

Tuple Fields
0: [u8; 32]
Implementations

impl MontgomeryPoint

pub fn mul_base(scalar: &Scalar) -> Self

Fixed-base scalar multiplication (i.e. multiplication by the base point).

pub const fn as_bytes(&self) -> &[u8; 32]

View this MontgomeryPoint as an array of bytes.

pub const fn to_bytes(&self) -> [u8; 32]

Convert this MontgomeryPoint to an array of bytes.

pub fn to_edwards(&self, sign: u8) -> Option<EdwardsPoint>

Attempt to convert to an EdwardsPoint, using the supplied choice of sign for the EdwardsPoint.

Inputs

sign: a u8 donating the desired sign of the resulting EdwardsPoint. 0 denotes positive and 1 negative.

Return

Some(EdwardsPoint) if self is the \(u\)-coordinate of a point on (the Montgomery form of) Curve25519;

None if self is the \(u\)-coordinate of a point on the twist of (the Montgomery form of) Curve25519;