Struct curve25519_dalek::ristretto::RistrettoBasepointTable

pub struct RistrettoBasepointTable(pub EdwardsBasepointTable);

Precomputation

Methods

impl RistrettoBasepointTable

pub fn create(basepoint: &RistrettoPoint) -> RistrettoBasepointTable

Create a precomputed table of multiples of the given basepoint.

pub fn basepoint(&self) -> RistrettoPoint

Get the basepoint for this table as a RistrettoPoint.