Struct curve25519_dalek::ristretto::RistrettoBasepointTable

pub struct RistrettoBasepointTable(/* private fields */);

A precomputed table of multiples of a basepoint, used to accelerate scalar multiplication.

A precomputed table of multiples of the Ristretto basepoint is available in the constants module:

use curve25519_dalek::constants::RISTRETTO_BASEPOINT_TABLE;
use curve25519_dalek::scalar::Scalar;

let a = Scalar::from(87329482u64);
let P = &a * RISTRETTO_BASEPOINT_TABLE;

Implementations

impl RistrettoBasepointTable

pub fn create(basepoint: &RistrettoPoint) -> RistrettoBasepointTable

Create a precomputed table of multiples of the given basepoint.

pub fn basepoint(&self) -> RistrettoPoint

Get the basepoint for this table as a RistrettoPoint.