Struct curve25519_dalek::scalar::Scalar

pub struct Scalar { /* fields omitted */ }

The Scalar struct represents an element in ℤ/lℤ, where

l = 2252 + 27742317777372353535851937790883648493

is the order of the basepoint. The Scalar is stored as bytes.

Methods

impl Scalar

pub fn from_bytes_mod_order(bytes: [u8; 32]) -> Scalar

Construct a Scalar by reducing a 256-bit integer modulo the group order.

pub fn from_canonical_bytes(bytes: [u8; 32]) -> Option<Scalar>

Attempt to construct a Scalar from a canonical byte representation.

Return

Some(s), where s is the Scalar corresponding to bytes, if bytes is a canonical byte representation;

None if bytes is not a canonical byte representation.

pub fn from_bits(bytes: [u8; 32]) -> Scalar

Construct a Scalar from the low 255 bits of a 256-bit integer.

This function is intended for applications like X25519 which require specific bit-patterns when performing scalar multiplication.

impl Scalar

pub fn random<T: Rng>(rng: &mut T) -> Self

Return a Scalar chosen uniformly at random using a user-provided RNG.

Inputs

rng: any RNG which implements the rand::Rng interface.

Returns

A random scalar within ℤ/lℤ.

pub fn hash_from_bytes<D>(input: &[u8]) -> Scalar where
D: Digest<OutputSize = U64> + Default,

Hash a slice of bytes into a scalar.

Takes a type parameter D, which is any Digest producing 64 bytes (512 bits) of output.

Convenience wrapper around from_hash.

Example

extern crate sha2;
use sha2::Sha512;

let msg = "To really appreciate architecture, you may even need to commit a murder";
let s = Scalar::hash_from_bytes::<Sha512>(msg.as_bytes());

pub fn from_hash<D>(hash: D) -> Scalar where
D: Digest<OutputSize = U64> + Default,

Construct a scalar from an existing Digest instance.

Use this instead of hash_from_bytes if it is more convenient to stream data into the Digest than to pass a single byte slice.

pub fn to_bytes(&self) -> [u8; 32]

Convert this Scalar to its underlying sequence of bytes.

pub fn as_bytes(&self) -> &[u8; 32]

View this Scalar as a sequence of bytes.

pub fn zero() -> Self

Construct the additive identity

pub fn one() -> Self

Construct the multiplicative identity

pub fn from_u64(x: u64) -> Scalar

Construct a scalar from the given u64.

pub fn invert(&self) -> Scalar

Compute the multiplicative inverse of this scalar.

pub fn multiply_add(a: &Scalar, b: &Scalar, c: &Scalar) -> Scalar

Compute (a * b) + c (mod l).

pub fn reduce(&self) -> Scalar

Reduce this Scalar mod l.

pub fn is_canonical(&self) -> bool

Check whether this Scalar is the canonical representative mod \(\ell\).

This is intended for uses like input validation, where variable-time code is acceptable.

// 2^255 - 1, since `from_bits` clears the high bit
let _2_255_minus_1 = Scalar::from_bits([0xff;32]);
assert!(!_2_255_minus_1.is_canonical());

let reduced = _2_255_minus_1.reduce();
assert!(reduced.is_canonical());

pub fn reduce_wide(input: &[u8; 64]) -> Scalar

Reduce a 512-bit little endian number mod l