Module curve25519_dalek::traits

Module for common traits.

Traits

BasepointTable

A precomputed table of basepoints, for optimising scalar multiplications.

Identity

Trait for getting the identity element of a point type.

IsIdentity

Trait for testing if a curve point is equivalent to the identity point.

MultiscalarMul

A trait for constant-time multiscalar multiplication without precomputation.

VartimeMultiscalarMul

A trait for variable-time multiscalar multiplication without precomputation.

VartimePrecomputedMultiscalarMul

A trait for variable-time multiscalar multiplication with precomputation.