Trait curve25519_dalek::traits::BasepointTable

pub trait BasepointTable {
    type Point;
    fn create(basepoint: &Self::Point) -> Self;

    fn basepoint(&self) -> Self::Point;

    fn basepoint_mul(&self, scalar: &Scalar) -> Self::Point;
}

A precomputed table of basepoints, for optimising scalar multiplications.

Associated Types

type Point

The type of point contained within this table.

Loading content...
Required methods

fn create(basepoint: &Self::Point) -> Self

Generate a new precomputed basepoint table from the given basepoint.

fn basepoint(&self) -> Self::Point

Retrieve the original basepoint from this table.

fn basepoint_mul(&self, scalar: &Scalar) -> Self::Point

Multiply a scalar by this precomputed basepoint table, in constant time.

Loading content...