Trait curve25519_dalek::traits::IsIdentity

pub trait IsIdentity {
    // Required method
    fn is_identity(&self) -> bool;
}

Trait for testing if a curve point is equivalent to the identity point.

Required Methods

fn is_identity(&self) -> bool

Return true if this element is the identity element of the curve.